+++ Assets/CampusSim/Tests/PlayMode/FixtureLifecycleTests.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections;
+using System.Linq;
 using InhaExpress.Client.Domain;
 using InhaExpress.Client.Networking;
 using InhaExpress.Client.Presentation;
@@ -7,6 +8,7 @@ using InhaExpress.Client.PC;
 using NUnit.Framework;
 using UnityEngine;
 using UnityEngine.TestTools;
+using UnityEngine.UI;
 
 namespace InhaExpress.Client.Tests
 {
@@ -17,7 +19,7 @@ namespace InhaExpress.Client.Tests
         {
             var hostObject = new GameObject("Test Host");
             var viewObject = new GameObject("Test View");
-            var source = new FixtureClientDataSource(ClientRole.PC_Operator, "0.1.4.0", sessionId: "lifecycle");
+            var source = new FixtureClientDataSource(ClientRole.PC_Operator, "0.1.5.0", sessionId: "lifecycle");
             try
             {
                 var host = hostObject.AddComponent<ClientRuntimeHost>();
@@ -30,6 +32,9 @@ namespace InhaExpress.Client.Tests
                 source.Start(0);
                 source.Pump(4);
                 StringAssert.Contains("ASSIGNED", view.Body);
+                Assert.That(view.GetComponentsInChildren<Canvas>(true).Length, Is.EqualTo(1));
+                var buttons = view.GetComponentsInChildren<Button>(true);
+                Assert.That(buttons.Select(x => x.name), Is.EquivalentTo(new[] { "Pause", "Restart", "Delivery" }));
                 view.enabled = false;
                 var frozen = view.Body;
                 source.Pump(14);